fix: 상세페이지 layout 삭제 및 page에 헤더 추가
